% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carson-package.R
\docType{package}
\name{carson-package}
\alias{carson}
\alias{carson-package}
\title{carson: Crosswalks Aligning Releases of SOC Or NAICS}
\description{
Provide crosswalks to simplify the temporal alignment of SOC and NAICS codes. SOC codes have been released in 2000, 2010, and 2018, and NAICS codes have been released every 5 years since 1997. BLS and Census (for SOC and NAICS, respectively) maintain reference data on code descriptions by release and concordances tables between successive releases. This package provides these tables as data objects, and automates conversions between years using the crosswalks.
}
\author{
\strong{Maintainer}: Aaron Weinstock \email{aaron.weinstock@nltgis.com}

}
\keyword{internal}
