% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_naics.R
\name{convert_naics}
\alias{convert_naics}
\title{Convert NAICS codes between years/releases}
\usage{
convert_naics(codes, from_year, to_year)
}
\arguments{
\item{codes}{A character vector of NAICS codes to be converted. All codes
must be valid NAICS codes (2 and 6 digits) present in the release
corresponding to \code{from_year} (see below).}

\item{from_year}{The year/release to convert from.}

\item{to_year}{The year/release to convert to.}
}
\value{
A data.table with 3 columns:
\itemize{
\item "code_\{from\}": the code in the release corresponding to \code{from_year}. This
will be a value given in \code{codes}.
\item "code_\{to\}": a possible value of "code_\{from\}" in the release
corresponding to \code{to_year}. There may be more than one "code_\{to\}" for each
"code_\{from\}", if "code_\{from\}" maps to multiple codes in the to-year
release.
\item "N": the number of 6-digit NAICS codes that fall under the umbrella of the
match. This can be used to make a rough estimate of how often the conversion
manifests. This is especially useful when one "code_\{from\}" matches to
multiple entries in "code_\{to\}".
}
}
\description{
Convert NAICS codes between years/releases
}
\details{
This function matches NAICS codes in a "from_year" to corresponding codes
in a "to_year". The from and to years are matched to releases of NAICS based
on the most current NAICS release during each year. New NAICS have been
published every 5 years from 1997 to 2022, so setting from_year = 2006 and
to_year = 2011 will, for example, convert 2002 NAICS to 2007 NAICS. All
codes to convert must be present in the NAICS release corresponding to the
from year. Any codes that are not found in the from-year release will be
identified in a warning before return.
}
\examples{
# Convert 3 codes from 2002 to 2022
convert_naics(c("32","451","926140"), 2002, 2022)
# Because of release schedule, the following is equivalent to the above
convert_naics(c("32","451","926140"), 2005, 2024)
}
