% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lance-package.R
\docType{package}
\name{lance-package}
\alias{lance}
\alias{lance-package}
\title{lance: Build NAICS Code Equivalences}
\description{
Identify equivalences of NAICS codes across various releases. New NAICS codes have been released every 5 years since 1997, and the US Census maintains records of code descriptions by year and crosswalk tables between successive years. This package automates lookups in those tables to build equivalences across different releases.
}
\author{
\strong{Maintainer}: Aaron Weinstock \email{aaron.weinstock@nltgis.com}

}
\keyword{internal}
