% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_2002}
\alias{naics_2002}
\title{NAICS 2002 2-6 digit codes}
\format{
\subsection{\code{naics_2002}}{

A data frame with 2,341 rows and 2 columns:
\describe{
\item{code}{2-6 digit 2002 NAICS code}
\item{description}{Description of the industry}
...
}
}
}
\source{
\url{https://www.census.gov/naics/reference_files_tools/2002/naics_2_6_02.txt}
}
\usage{
naics_2002
}
\description{
2-6 digit industry codes and descriptions for NAICS 2002.
}
\keyword{datasets}
