% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_2007}
\alias{naics_2007}
\title{NAICS 2007 2-6 digit codes}
\format{
\subsection{\code{naics_2007}}{

A data frame with 2,328 rows and 2 columns:
\describe{
\item{code}{2-6 digit 2007 NAICS code}
\item{description}{Description of the industry}
...
}
}
}
\source{
\url{https://www.census.gov/naics/reference_files_tools/2007/naics07.xls}
}
\usage{
naics_2007
}
\description{
2-6 digit industry codes and descriptions for NAICS 2007.
}
\keyword{datasets}
