% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_2012}
\alias{naics_2012}
\title{NAICS 2012 2-6 digit codes}
\format{
\subsection{\code{naics_2012}}{

A data frame with 2,209 rows and 2 columns:
\describe{
\item{code}{2-6 digit 2012 NAICS code}
\item{description}{Description of the industry}
...
}
}
}
\source{
\url{https://www.census.gov/naics/2012NAICS/2-digit_2012_Codes.xls}
}
\usage{
naics_2012
}
\description{
2-6 digit industry codes and descriptions for NAICS 2012.
}
\keyword{datasets}
