% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_2017}
\alias{naics_2017}
\title{NAICS 2017 2-6 digit codes}
\format{
\subsection{\code{naics_2017}}{

A data frame with 2,196 rows and 2 columns:
\describe{
\item{code}{2-6 digit 2017 NAICS code}
\item{description}{Description of the industry}
...
}
}
}
\source{
\url{https://www.census.gov/naics/2017NAICS/2-6\%20digit_2017_Codes.xlsx}
}
\usage{
naics_2017
}
\description{
2-6 digit industry codes and descriptions for NAICS 2017.
}
\keyword{datasets}
