% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_2022}
\alias{naics_2022}
\title{NAICS 2022 2-6 digit codes}
\format{
\subsection{\code{naics_2022}}{

A data frame with 2,125 rows and 2 columns:
\describe{
\item{code}{2-6 digit 2022 NAICS code}
\item{description}{Description of the industry}
...
}
}
}
\source{
\url{https://www.census.gov/naics/2022NAICS/2-6\%20digit_2022_Codes.xlsx}
}
\usage{
naics_2022
}
\description{
2-6 digit industry codes and descriptions for NAICS 2022.
}
\keyword{datasets}
