% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_xwalk}
\alias{naics_xwalk}
\title{NAICS 1997-2022 6-digit code crosswalk}
\format{
\subsection{\code{naics_xwalk}}{

A data frame with 1,615 rows and 6 columns:
\describe{
\item{code_YYYY}{6-digit NAICS code in year YYYY}
...
}
}
}
\source{
Concordance tables available at \url{https://www.census.gov/naics/?48967}
}
\usage{
naics_xwalk
}
\description{
Full crosswalk for 6-digit codes for all NAICS releases.
}
\keyword{datasets}
