% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc_2000}
\alias{soc_2000}
\title{SOC 2000 codes}
\format{
\subsection{\code{soc_2000}}{

A data frame with 821 rows and 3 columns:
\describe{
\item{code}{2000 SOC code}
\item{title}{Name of the occupation}
\item{definition}{Description of the occupation}
...
}
}
}
\source{
\url{https://www.bls.gov/soc/2000/soc-definitions-2000.xlsx}
}
\usage{
soc_2000
}
\description{
Occupation codes and definitions for SOC 2000.
}
\keyword{datasets}
