% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc_2010}
\alias{soc_2010}
\title{SOC 2010 codes}
\format{
\subsection{\code{soc_2010}}{

A data frame with 840 rows and 3 columns:
\describe{
\item{code}{2010 SOC code}
\item{title}{Name of the occupation}
\item{definition}{Description of the occupation}
...
}
}
}
\source{
\url{https://www.bls.gov/soc/soc_2010_definitions.xls}
}
\usage{
soc_2010
}
\description{
Occupation codes and definitions for SOC 2010.
}
\keyword{datasets}
