% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc_2018}
\alias{soc_2018}
\title{SOC 2018 codes}
\format{
\subsection{\code{soc_2018}}{

A data frame with 1,447 rows and 4 columns:
\describe{
\item{group}{Level of occupation (major, minor, broad, detailed)}
\item{code}{2018 SOC code}
\item{title}{Name of the occupation}
\item{definition}{Description of the occupation}
...
}
}
}
\source{
\url{https://www.bls.gov/soc/2018/soc_2018_definitions.xlsx}
}
\usage{
soc_2018
}
\description{
Occupation codes and definitions for SOC 2018.
}
\keyword{datasets}
