% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc_xwalk}
\alias{soc_xwalk}
\title{SOC 2000-2018 code crosswalk}
\format{
\subsection{\code{soc_xwalk}}{

A data frame with 920 rows and 3 columns:
\describe{
\item{code_YYYY}{SOC code in year YYYY}
...
}
}
}
\source{
Crosswalk tables available at \url{https://www.bls.gov/soc}
}
\usage{
soc_xwalk
}
\description{
Full crosswalk for codes for all SOC releases.
}
\keyword{datasets}
